requirements.txt
# Core LLM Provider Libraries
openai>=1.0.0
anthropic>=0.5.0
ibm-watson-machine-learning>=1.0.0
requests>=2.28.0  # For Ollama client

# Utilities
python-dotenv>=0.19.0  # For environment variable loading

# Testing (development only)
pytest>=7.0.0
pytest-cov>=3.0.0